+++ resources/views/livewire/attendance/poll-edit.blade.php
@@ -14,7 +14,18 @@ addEvents() {
     $wire.addEventsToSelection(this.additionalEventList);
     this.additionalEventList = [];
 }}">
-    <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg mb-5 p-5 flex justify-end gap-2 items-center">
+    <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg mb-5 p-5 flex justify-between gap-2 items-center">
+        <x-default-button
+            x-data="{ clickCnt: 0, onClick() {
+                        if(this.clickCnt > 0) {
+                            $wire.deletePoll();
+                        } else {
+                            this.clickCnt++;
+                            $el.innerHTML = 'Are you sure?';
+                        }
+                    }}"
+            x-on:click="onClick()" title="Delete this poll"
+            class="btn-danger">{{ __('Delete poll') }}</x-default-button>
         <div class="flex gap-2 items-center">
             <span x-cloak class="text-gray-500 text-xs mt-1"
                   x-show="additionalEventList.length > 0">Add selected events or unselect them.</span>
@@ -30,7 +41,7 @@ addEvents() {
     <div class="flex flex-col lg:grid lg:grid-cols-2 gap-4">
 
         <div class="bg-white shadow-sm sm:rounded-lg p-4">
-           <x-livewire.attendance.poll-basics />
+            <x-livewire.attendance.poll-basics/>
         </div>
         <div class="bg-white shadow-sm sm:rounded-lg p-4">
             <x-livewire.attendance.poll-event-selection :pollForm="$pollForm"/>