+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Changed
 - change event filter date persistence to cookieStorage
+- redesigned notification messages
+- update message adding
+### Added
+- Delete button in poll edit form
 
 ## v0.6.2 - 2024-03-07
 ### Added